refactor codes (#73)
* fix: refactor pyproject.toml file, delete setup.cfg

* fix: refactor pyproject.toml file, delete setup.cfg

* fix: refactor pyproject.toml file, delete setup.cfg
