+++ pyproject.toml
@@ -1,3 +1,47 @@
 [build-system]
 requires = ["setuptools>=68.1.2"]
 build-backend = "setuptools.build_meta"
+
+[project]
+name = "scar"
+version = "0.5.5"
+requires-python = ">= 3.10"
+dependencies = [
+    "torch >= 1.10.0",
+    "torchvision >= 0.9.0",
+    "tqdm >= 4.62.3",
+    "seaborn >= 0.11.2",
+    "scikit-learn >= 1.0.1",
+    "pyro-ppl >= 1.8.0",
+    "scanpy >= 1.9.2"
+]
+authors = [
+    {name = "Caibin Sheng", email = "caibin.sheng.res@gmail.com"}
+]
+description = "scAR (single-cell Ambient Remover) is a package for denoising the ambient signals in droplet-based single cell omics"
+readme = "README.md"
+license = {text = "MIT License"}
+keywords = ["single cell omics", "variational autoencoder", "machine learning", "generative model", "cite-seq", "scCRISPRseq", "scRNAseq"]
+
+[project.urls]
+Homepage = "https://github.com/Novartis/scAR"
+Documentation = "https://scar-tutorials.readthedocs.io/en/main/"
+Repository = "https://github.com/Novartis/scar.git"
+Issues = "https://github.com/Novartis/scAR/issues"
+Changelog = "https://github.com/me/spam/blob/master/CHANGELOG.md"
+
+[tool.semantic_release]
+version_toml = ["pyproject.toml:project.version"]
+major_on_zero = false
+branch = "develop"
+upload_to_release = false
+hvcs = "github"
+upload_to_repository = false
+upload_to_pypi = false
+patch_without_tag = false
+
+[tool.semantic_release.changelog]
+changelog_file="docs/Release_notes.md"
+
+[project.gui-scripts]
+scar = "scar.main.__main__:main"
