+++ scar/__init__.py
@@ -1,6 +1,7 @@
 # -*- coding: utf-8 -*-
+from importlib.metadata import version
+__version__ = version("scar")
 
-from .main.__version__ import __version__
 from .main._scar import model
 from .main._setup import setup_anndata
 from .main import _data_generater as data_generator
