--- scar/main/__version__.py
@@ -1,6 +1,0 @@
-# -*- coding: utf-8 -*-
-"""version here"""
-
-__version__ = "0.5.5"
-_author = "Caibin Sheng"
-_copyright = "2022, Novartis Institute for BioMedical Research"
