--- setup.cfg
@@ -1,45 +1,0 @@
-[metadata]
-name = scar
-version = attr: scar.main.__version__.__version__
-author = Caibin Sheng
-author_email = caibin.sheng.res@gmail.com
-url = https://github.com/Novartis/scAR
-description = scAR (single-cell Ambient Remover) is a package for denoising the ambient signals in droplet-based single cell omics
-long_description = file: README.md
-long_description_content_type = text/markdown
-keywords = single cell omics, variational autoencoder, machine learning, generative model, cite-seq, scCRISPRseq, scRNAseq
-license = MIT
-classifiers =
-    License :: OSI Approved :: MIT License
-    Programming Language :: Python :: 3
-    Topic :: Scientific :: Artificial Intelligence
-
-[options]
-packages = find:
-python_requires = >=3.10
-zip_safe = False
-include_package_data = True
-install_requires =
-    torch >= 1.10.0
-    torchvision >= 0.9.0
-    tqdm >= 4.62.3
-    seaborn >= 0.11.2
-    scikit-learn >= 1.0.1
-    pyro-ppl >= 1.8.0
-    scanpy >= 1.9.2
-package_dir=
-    =.
-
-[options.entry_points]
-console_scripts =
-    scar = scar.main.__main__:main
-
-[semantic_release]
-version_variable=scar/main/__version__.py:__version__
-branch=develop
-hvcs=github
-upload_to_repository=false
-upload_to_pypi=false
-upload_to_release=false
-patch_without_tag=false
-changelog_file=docs/Release_notes.md
